indexer-node/src/requirements.txt
boto3
python-dateutil
base58
git+https://github.com/tehmaze/python-multihash.git
# git+https://github.com/koriaf/python-goose.git
lxml

gevent
httplib2
beautifulsoup4
